v3.0.1

+++ utils/tensorboard_parser.py
@@ -5,17 +5,19 @@ import pandas as pd
 import os
 
 class tensorboard_parser:
-    def __init__(self, log_path):
+    def __init__(self, log_path, save=False, plot=False):
         self.log_path = log_path
         self.ea = event_accumulator.EventAccumulator(log_path)
         self.ea.Reload()
         self.scalar_keys = self.ea.scalars.Keys()
         self.stats_df = pd.DataFrame(columns=['Metric', 'Mean', 'Max', 'Min', 'Std Dev', 'Variance'])
-        # Remove the plot directory if it exists
-        if os.path.exists('./plot'):
-            import shutil
-            shutil.rmtree('./plot')
-        os.makedirs('./plot')
+        self.save = save
+        self.plot = plot
+        if self.save:
+            if os.path.exists('./plot'):
+                import shutil
+                shutil.rmtree('./plot')
+            os.makedirs('./plot')
 
     # Function to compute descriptive statistics
     def compute_statistics(self, values):
@@ -27,7 +29,12 @@ class tensorboard_parser:
         return mean, max_val, min_val, std_dev, variance
 
     def parse_and_plot(self):
-        for key in self.scalar_keys:
+        num_metrics = len(self.scalar_keys)
+        fig, axes = plt.subplots(num_metrics, 1, figsize=(15, 5 * num_metrics))
+        if num_metrics == 1:
+            axes = [axes]  # Ensure axes is a list even with one plot
+        
+        for idx, key in enumerate(self.scalar_keys):
             # Retrieve scalar events for the key
             events = self.ea.Scalars(key)
             
@@ -48,17 +55,21 @@ class tensorboard_parser:
                 'Variance': variance
             }, ignore_index=True)
             
-            # Create a plot for the scalar metric
-            plt.figure()
-            plt.plot(steps, values)
-            plt.title(f'{key}\nMean: {mean:.4f}, Max: {max_val:.4f}, Min: {min_val:.4f}, Std Dev: {std_dev:.4f}, Variance: {variance:.4f}')
-            plt.xlabel('Steps')
-            plt.ylabel('Value')
-            plt.grid(True)
-            # Make sure the plot directory exists
-            os.makedirs(f"./plot/{key.split('/')[0]}", exist_ok=True)
-            plt.savefig(f'./plot/{key}.png',dpi=300)
-            # Show the plot
+            # Plot the scalar metric
+            axes[idx].plot(steps, values)
+            axes[idx].set_title(f'{key}\nMean: {mean:.4f}, Max: {max_val:.4f}, Min: {min_val:.4f}, Std Dev: {std_dev:.4f}, Variance: {variance:.4f}')
+            axes[idx].set_xlabel('Steps')
+            axes[idx].set_ylabel('Value')
+            axes[idx].grid(True)
+        
+        plt.tight_layout()
+        
+        if self.save:
+            os.makedirs("./plot", exist_ok=True)
+            plt.savefig('./plot/all_metrics.png', dpi=300)
+        
+        # Show the plot
+        if self.plot:
             plt.show()
 
         # Display the statistics DataFrame
@@ -77,10 +88,13 @@ class tensorboard_parser:
             
             # Compute statistics
             mean, max_val, min_val, std_dev, variance = self.compute_statistics(values)
-            sample_condition = np.mod(np.array(values).__index__(), 50) == 0
+            
+            # Sample every 50th value
+            sample_condition = np.arange(len(values)) % 50 == 0
             sample = np.array(values)[sample_condition]
+            
             # Append statistics to the DataFrame
-            self.stats_df = self.stats_df.append({
+            self.stats_df = self.stats_df._append({
                 'Metric': key,
                 'Mean': mean,
                 'Max': max_val,
@@ -99,7 +113,8 @@ class tensorboard_parser:
     
 if __name__ == '__main__':
     # Create an instance of the tensorboard_parser class
-    tb_parser = tensorboard_parser('runs/ppo_1')
+    path = r"H:\Omniverse\Library\isaac_sim-2023.1.1\OmniIsaacGymEnvs\omniisaacgymenvs\runs\Crazyflie\summaries\events.out.tfevents.1715672304.DESKTOP-P1IIRLN"
+    tb_parser = tensorboard_parser(path,save=True, plot=True)
     # Parse the log files and plot the scalar metrics
     tb_parser.parse_and_plot()
     # Parse the log files and return the DataFrame
